# Write images in buffers