@openrm/dev 

This is an attempt to register/write the image streams as they are, whenever possible.
It is done by replacing the existing redundant memory copies (full reads) with byte-wise operations, and instead doing it only once for adding the `Length` metadata on each object.