@PuKoren 2021-01-18
thanks! also I would be interested to know about the performances gains, if it does not take you too much time