@naggingant 2021-01-19
Quickly ran benchmarks, and basically it is not always "better", and I remembered this was the primary reason I did not open this PR first :sweat_smile: 

The collected benchmark results and profiles were quite hard to understand at the moment, but I'll try to summarize them here.
First, the CPU time is longer for most of <1mb images, since the zlib de/compression works much faster with data all in memory (than data coming in stream)
And even the memory footprint did get worse for some cases (<1mb images mostly), and it could be a bug or something I cannot tell right now. I didn't get anything yet from the `alloc_space` profiles, which claim the branch performs better than the original.

Attaching the benchmark results below :pray: 

i) ~900kb PNG
```
(original)
BenchmarkPNGEmbedding-12              13          86746710 ns/op        58313659 B/op       4612 allocs/op
(this branch)
BenchmarkPNGEmbedding-12               3         479347188 ns/op        127033541 B/op      4407 allocs/op
(this branch + sync.Pool)
BenchmarkPNGEmbedding-12               3         461659600 ns/op        73372589 B/op       4370 allocs/op
```

ii) ~2mb PNG
```
(original)
BenchmarkPNGEmbedding-12             128           9228263 ns/op        39116499 B/op        900 allocs/op
(this branch)
BenchmarkPNGEmbedding-12             136           8186439 ns/op        16689900 B/op        544 allocs/op
(this branch + sync.Pool)
BenchmarkPNGEmbedding-12             368           3261043 ns/op         1216504 B/op        498 allocs/op
```