@PuKoren 2021-01-19
the gains on 2mb + sync.Pool are impressive

the allocs/op on the 900kb image is quite high, do you know why there is much more allocs than for the 2mb png?