@naggingant 2021-01-19
Yeah that's really weird, maybe there is some difference in the format (PNG can contain many things)
Now I suspect I did something wrong with the benchmarks, or there is a bug...
Let me get back to you later

This hobby work does not prevent the other development :+1: